+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-None.
+### Added
+
+* [[#63](https://github.com/MyWarp/MyWarp/issues/63)] - Add a configuration option to specify which
+  warps are shown on Dynmap
 
 ## [3.1.1] - 2021-01-12